requirements.txt
pyrogram 
tgcrypto
requests
python-dotenv